Add support for attached motion sensor on GPIO
This assumes running on Raspberry Pi, using an HC-SR501. This device
outputs a HIGH signal when motion is detected, or a low signal when
no motion is detected after a certain delay.

Can be set by specifying --motion-pin to the GPIO pin on the raspberry
to which the HC-SR501 data pin is connected.

Implements #5